Fixed a bug in our classpath Removed MySQL jar. Renamed the database api for sqlite to sql.
