+++ src/database/DatabaseInterface.java
@@ -3,4 +3,7 @@ package database;
 
 public interface DatabaseInterface {
 
+	static final String DBPATH="data/xml/";
+	
+	
 }
