+++ src/database/sql/Database.java
@@ -1,4 +1,4 @@
-package database.sqlite;
+package database.sql;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
